Review fixes:
- Reduce line lengths of tests to avoid hiding assertions
- Fix style for one line if
- Rename ProgramResult.apply to avoid the impression it's double wrapping
- Remove `anyCommandFailed` field and call `hasAnyCommandFailed` on demand instead
  - Also move `hasAnyCommandFailed` to UI as it is only used at the end to determine the exit code
