+++ src/main/scala/com/gu/ssm/Logic.scala
@@ -80,8 +80,4 @@ object Logic {
     }
   }
 
-  def hasAnyCommandFailed(ssmResults: List[(InstanceId, Either[CommandStatus, CommandResult])]): Boolean = {
-    ssmResults.exists { case(_, result) => result.exists(_.commandFailed) }
-  }
-
 }
